+++ purchasely/ios/Classes/SwiftPurchaselyFlutterPlugin.swift
@@ -148,14 +148,15 @@ public class SwiftPurchaselyFlutterPlugin: NSObject, FlutterPlugin {
             closePresentation()
         case "signPromotionalOffer":
             signPromotionalOffer(arguments: arguments, result: result)
+        case "isEligibleForIntroOffer":
+            isEligibleForIntroOffer(arguments: arguments, result: result)
         default:
             result(FlutterMethodNotImplemented)
         }
     }
 
     private func isAnonymous(result: @escaping FlutterResult) {
-        let returnValue: [String: Any] = ["result": Purchasely.isAnonymous()]
-        result(returnValue)
+        result(Purchasely.isAnonymous())
     }
 
     private func hidePresentation() {
@@ -182,6 +183,23 @@ public class SwiftPurchaselyFlutterPlugin: NSObject, FlutterPlugin {
             }
         }
     }
+    
+    private func isEligibleForIntroOffer(arguments: [String: Any]?, result: @escaping FlutterResult) {
+        guard let arguments = arguments, let planVendorId = arguments["planVendorId"] as? String else {
+            result(FlutterError.failedArgumentField("planVendorId", type: String.self))
+            return
+        }
+        
+        DispatchQueue.main.async {
+            Purchasely.plan(with: planVendorId) { plan in
+                plan.isUserEligibleForIntroductoryOffer { res in
+                    result(res)
+                }
+            } failure: { error in
+                result(FlutterError.error(code:"-1", message:"plan \(planVendorId) not found", error: error))
+            }
+        }
+    }
 
     private func start(arguments: [String: Any]?, result: @escaping FlutterResult) {
 
@@ -190,7 +208,7 @@ public class SwiftPurchaselyFlutterPlugin: NSObject, FlutterPlugin {
             return
         }
 
-		Purchasely.setSdkBridgeVersion("4.0.1")
+		Purchasely.setSdkBridgeVersion("4.1.0")
         Purchasely.setAppTechnology(PLYAppTechnology.flutter)
 
         let logLevel = PLYLogger.LogLevel(rawValue: (arguments["logLevel"] as? Int) ?? PLYLogger.LogLevel.debug.rawValue) ?? PLYLogger.LogLevel.debug
@@ -221,7 +239,7 @@ public class SwiftPurchaselyFlutterPlugin: NSObject, FlutterPlugin {
             return
         }
 
-		Purchasely.setSdkBridgeVersion("4.0.1")
+		Purchasely.setSdkBridgeVersion("4.1.0")
         Purchasely.setAppTechnology(PLYAppTechnology.flutter)
 
         let logLevel = PLYLogger.LogLevel(rawValue: (arguments["logLevel"] as? Int) ?? PLYLogger.LogLevel.debug.rawValue) ?? PLYLogger.LogLevel.debug